feat: Make mermaid in electron simpler due to being ran in renderer already
