--- packages/mermaid-electron-renderer/src/mermaid_renderer.js
@@ -1,15 +1,0 @@
-import mermaid from "mermaid";
-
-window.renderMermaidChart = async (chartData, mermaidConfig) => {
-	mermaid.initialize({ ...mermaidConfig, startOnLoad: false });
-
-	const { svg } = await mermaid.render("graphDiv2", chartData);
-	const chartElement = document.querySelector("#graphDiv");
-	chartElement.innerHTML = svg;
-
-	const svgElement = document.querySelector("#graphDiv svg");
-	return {
-		width: svgElement.scrollWidth,
-		height: svgElement.scrollHeight,
-	};
-};
